requirements.txt
dj-database-url==0.4.1
Django==2.0
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.2

social-auth-app-django
django-bootstrap3

djangorestframework
django-cors-headers
django-oauth-toolkit

